==> workbook/01_data_review_data_cleaning.ipynb <==
# Data review: 

- check out the data structure
- selected the relevant columns
- re-save the data with a proper separater for Tableau to read

import pandas as pd
import datetime
import numpy as np

df = pd.read_csv("/Users/Tianwang/ironhack/git/03_project_ironhack_tableau/data/raw_data_Animal_Control_Incidents.csv")

df

# Deduplication 

df_unique = df.drop_duplicates(keep=False,inplace=False) 

len(df_unique)

len(df)

df = df_unique.copy()

df.columns

#selecting columnes of interest.

df = df[['FILE NUMBER', 'INCIDENT DATE', 'INCIDENT TIME', 'DISPATCHED TIME','DISPATCHED SITUATION',
       'ARRIVAL TIME', 'AVAILABLE TIME', 'COMPLAINANT STREET ', 'REQUEST TYPE',
       'OFFICER', 'EQUIPMENT NUMBER', 'REMARKS', 'IMPOUND NUMBER',
       'IMPOUND DATE', 'IMPOUND TIME', 'LOCATION', 'MUNICIPALITY ', 'ZIP CODE',
       'SERVICE CODE', 'SPECIES', 'BREED', 'SEX', 'SIZE', 'COLOR', 'AGE',
       'CONDITION', 'TEMPERMENT', 'COLLAR', 'COLLAR COLOR','PET NAME', 'DISPOSITION',
       'DISPOSITION DATE', 'DISPOSITION OFFICER']]

# Clean up information around date and time

1. There are empty fields. "INCIDENT TIME" 

df["INCIDENT TIME"].value_counts(dropna=False)

df.loc[df["INCIDENT TIME"]=="00:00"] 

2. However, there are 4715 rows containing DISPATCHED TIME to be 00:00. / And 10513 NaNs. 

df["DISPATCHED TIME"].value_counts(dropna=False)

df.loc[df["DISPATCHED TIME"]=="00:00"] 

And 5930 rows containing ARRIVAL TIME to be 00:00. // 11217 NaN

df["ARRIVAL TIME"].value_counts(dropna=False)

df.loc[df["ARRIVAL TIME"]=="00:00"] 

df = df.assign(arrival_time_new = pd.to_datetime(df.loc[:,"INCIDENT DATE"] + " " + df.loc[: , "ARRIVAL TIME"]))

df["year"] = df["arrival_time_new"].dt.year.astype(str).str.strip(".0")

#manually changed the record for filenumber "201634939": Available Time from 01:83 to 18:30.

df.loc[df["AVAILABLE TIME"]=="01:83",["AVAILABLE TIME"]]="18:30"

df = df.assign(available_time_new = pd.to_datetime(df["INCIDENT DATE"]+ " "+df["AVAILABLE TIME"]))

df = df.assign(incident_time_new = pd.to_datetime(df["INCIDENT DATE"]+ " "+df["INCIDENT TIME"]))

#manually changed the record for filenumber "201629609": Dispatched Time from ":" to Indicdent Time "09:09"

df.loc[df["DISPATCHED TIME"]==":",["DISPATCHED TIME"]]="09:09"

df = df.assign(dispatched_time_new = pd.to_datetime(df["INCIDENT DATE"]+ " "+df["DISPATCHED TIME"]))

# Calculate the time bewteen incident, dispatching and arrival - only if all 3 dates are available. 

Create a new columns containing the information whether all 3 types of data are available (incident, dispatching and arrival). If yes, the column contains "Yes", otherwise "No". This column can be used as a Tableau filter wenn doing calculations on those time measures.

df = df.assign(all_data_available = "Yes")

df.loc[(df["ARRIVAL TIME"] == "00:00") | (df["ARRIVAL TIME"].isna()),["all_data_available"]] = "No"
df.loc[(df["DISPATCHED TIME"] == "00:00") | (df["DISPATCHED TIME"].isna()),["all_data_available"]] = "No"
df.loc[(df["INCIDENT TIME"] == "00:00") | (df["INCIDENT TIME"].isna()),["all_data_available"]] = "No"

First testing out the method on simple examples using datetime. However, I then decided to convert all new columns into datetime, so don't need to use the strptime method anymore. 

import datetime

dt_string = "12/27/2016 16:17"
dt_string_1 = "09/11/2019 14:08"
test = datetime.datetime.strptime(dt_string, "%m/%d/%Y %H:%M")
test_1 = datetime.datetime.strptime(dt_string_1, "%m/%d/%Y %H:%M")

test_1

Diff = (test_1 - test)

Diff

Calculate the time different from in minutes.
    1) Incident - Dispatch
    2) Dispatch - Arrival

df = df.assign(incident_to_dispatch = ((df["dispatched_time_new"]-df["incident_time_new"]).dt.seconds/60))
df = df.assign(disp_to_arrival = ((df["arrival_time_new"] - df["dispatched_time_new"]).dt.seconds/60))

# df.to_csv("/Users/Tianwang/ironhack/git/03_project_ironhack_tableau/data/Baton_Rouge_Animal_incident_new.csv", sep=";", index=False)

# Check out the data quality on "Request Type" and "Service Code"

df["REQUEST TYPE"].value_counts()

df["SERVICE CODE"].value_counts()

--> It seems that the Column "Service Code" contains more clean data. 

Put categories together, which only differ from the way of writing.

df.loc[(df["SERVICE CODE"]=="STRAY AN SECURED") | (df["SERVICE CODE"]=="STRAY AN. SECURED"), ["SERVICE CODE"]] = "STRAY AND SECURED"

df.loc[(df["SERVICE CODE"]=="DANGEROUS ANIMAL SIT.") | (df["SERVICE CODE"]=="DANGEROUS ANIMAL SIT"), ["SERVICE CODE"]] = "DANGEROUS ANIMAL SIT"

df["SERVICE CODE"].value_counts()

As there are several so many subcategories, I decided to create a new columns having a more aggregated view. 

df["service_code_summary"] = df["SERVICE CODE"]

# STRAY: STRAY and STRAY AN SECURED
df.loc[(df["SERVICE CODE"] == "STRAY") | (df["SERVICE CODE"] == "STRAY AND SECURED"),["service_code_summary"]] = "STRAY"

# BITE CASE: 
df.loc[(df["SERVICE CODE"] == "BITE CASE")  | (df["SERVICE CODE"] == "BITE CASE SECURE")|(df["SERVICE CODE"] == "BITE CASE LOOSE"),["service_code_summary"]] = "BITE CASE"

# PUT "HIT BY CAR" to "INJURED"
df.loc[(df["SERVICE CODE"] == "HIT BY CAR"), ["service_code_summary"]] = "INJURED"

# COMBINE"VET PICKUP" AND "SICK ANIMAL"
df.loc[(df["SERVICE CODE"] == "VET PICKUP"), ["service_code_summary"]] = "SICK ANIMAL"

# COMBINE "LOOSE LIVESTOCK" TO "LOOSE"
df.loc[(df["SERVICE CODE"] == "LOOSE LIVESTOCK"), ["service_code_summary"]] = "LOOSE"

# ADD "SNAKE" TO "DANGEROUS ANIMAL SIT"
df.loc[(df["SERVICE CODE"] == "SNAKE"), ["service_code_summary"]] = "DANGEROUS ANIMAL SIT"

# COMBINE "SPECIAL CASE", "UNKNOW", "SCHOOL CALL" to "OTHER"
df.loc[(df["SERVICE CODE"] == "SPECIAL CASE"), ["service_code_summary"]] = "OTHER"
df.loc[(df["SERVICE CODE"] == "UNKNOWN"), ["service_code_summary"]] = "OTHER"
df.loc[(df["SERVICE CODE"] == "SCHOOL CALL"), ["service_code_summary"]] = "OTHER"
df.loc[(df["SERVICE CODE"] == "OWNED"), ["service_code_summary"]] = "OTHER"

df["service_code_summary"].value_counts()

df[["SERVICE CODE","service_code_summary"]]

# Check out the data quality on SPECIES', 'BREED'

Remark: NaN is not included in the value_counts.

df["SPECIES"].value_counts()

#combine UNKNOWN and OTHER
df["species_clean"] = df["SPECIES"]
df.loc[(df["SPECIES"] == "UNKNOWN"), ["species_clean"]] = "OTHER"

#cobine BIRD and BIRDS
df.loc[(df["SPECIES"] == "BIRDS"), ["species_clean"]] = "BIRD"

df["species_clean"].value_counts()

int(len(df["species_clean"])) - int(df["species_clean"].count()) #number of NaN values 

dog_breed = pd.DataFrame(df["BREED"].where(df["SPECIES"]=="DOG").value_counts())

dog_breed

--> Will clean the Breed data for dogs, as I want to use it for visualization. 

# Clean the adress / get the geo Coordinates thru API 

Found a page https://opencagedata.com/api, which translate adress into Geo coordinates. It will require some time to do. In case there is time for it, I want to add the information to the Tableau board. 

df["clean_adress"] = ""

df["ZIP CODE"] = df["ZIP CODE"].astype(str).str.zfill(5)

df["ZIP CODE"] = df["ZIP CODE"].str.strip(".0")

df["clean_adress"] = df["LOCATION"] + ", " + df["MUNICIPALITY "] + ", " + df["ZIP CODE"]

df["clean_adress"]

# Check out the data quality on Temperment & Condition

df["TEMPERMENT"].count() 

len(df) #roughly 20.000 entries without a "TEMPERMENT"

df[df["TEMPERMENT"].isna()]["year"].value_counts()

df["CONDITION"].count() 

df[df["CONDITION"].isna()]["year"].value_counts()

# File number with multiple entries

df["FILE NUMBER"].value_counts()

df.loc[df["FILE NUMBER"]==2014274530]

--> I cross-checked a few. My hypothesis: 1 file number is for 1 incident, but in case multiple animals are involved, each animal gets a separate row. 

# Select the clean columns and convert the file to a csv

select_columns = ['FILE NUMBER', 'INCIDENT DATE','COMPLAINANT STREET ','LOCATION',
       'MUNICIPALITY ', 'ZIP CODE', 'SERVICE CODE', 'SPECIES', 'BREED','CONDITION', 'TEMPERMENT','arrival_time_new', 'available_time_new',
       'incident_time_new', 'dispatched_time_new', 'all_data_available',
       'incident_to_dispatch', 'disp_to_arrival', 'service_code_summary',
       'species_clean','clean_adress']

df_select = df[select_columns]

new_names = ["file_number","incident_date","complainant_street","location","municipality",
             "zip_code","service_code","species","breed","condition","temperment","arrival_time_new",
             "available_time_new","incident_time_new", "dispatched_time_new", "all_time_available",
             "incident_to_dispatch", "disp_to_arrival", "service_code_summary", "species_clean","clean_adress"]

dictionary = dict(zip(select_columns, new_names))

df_select = df_select.rename(columns = dictionary)
df_select

pd.to_pickle(df_select, "/Users/Tianwang/ironhack/git/03_project_ironhack_tableau/data/baton_rouge_animal_control_incidents_clean.pkl")

df_select.to_csv("/Users/Tianwang/ironhack/git/03_project_ironhack_tableau/data/baton_rouge_animal_control_incidents_clean.csv", sep=";", index=False)

pd.to_pickle(df, "/Users/Tianwang/ironhack/git/03_project_ironhack_tableau/data/baton_rouge_animal_control_incidents_deep_dive.pkl")
