==> workbook/02. deep_dive_missing_data.ipynb <==
# 2. Deep_dive_missing_data

When first plotting the data in Tableau, one can observe: 
1) for year 2016-2018, the number of cases are as doubled compared to the preious years and 2019. 
2) for year 2016-2018, almost half of the records didn't contain complete information, especially for SERVICE CODE, location, species, breed etc.

The hypothesis is:
--->REQUEST CODE and COMPLAINANT_STREET contains information collected during the emergency phone call. 

--> However, only if the Baton Rouge Animal Control and Rescue Center (ACRC) provided service ("SERVICE CODE"), more detailed information such as breed, temperment, condition, species and location are collected. 

--> Empty SERVICE CODE fields only appeared in the years 2016-2018 and the systems has been changed back to only recordings cases, where service has been provided. 

--> Therefore, as a consequence, I decided to take out all records in the years 2016-2019, whenever the "SERVICE CODE" is not filled out. 

To test my hypothesis, I used the table with all relevant columns and did several cross-checks. 

import pandas as pd

df = pd.read_pickle("/Users/Tianwang/ironhack/git/03_project_ironhack_tableau/data/baton_rouge_animal_control_incidents_deep_dive.pkl")

#SERVICE CODE = NaN / Hypothesis: REQUEST TYPE IS NOT empty. --> proven

df[df["SERVICE CODE"].isnull()]["REQUEST TYPE"].value_counts(dropna = False)

df["SERVICE CODE"].value_counts(dropna = False) 

df["SPECIES"].value_counts(dropna = False) #hypothesis: if Species = NaN, Service Code is empty as well.

df[df["SPECIES"].isnull()]["SERVICE CODE"].value_counts(dropna = False) #hypothesis proven. 

df["SERVICE CODE"].value_counts(dropna = False) 
#hypothesis: if SERVICE CODE is NaN, all other relevant fields are NaN as well.(e.g. Breed, Species, Temperment, Condition)

test = df.loc[df["SERVICE CODE"].isnull(),["SPECIES","TEMPERMENT","CONDITION"]]

test["SPECIES"].value_counts(dropna=False)

test["TEMPERMENT"].value_counts(dropna=False)

test["CONDITION"].value_counts(dropna=False)

# Remove records, where no service has been provided.

df_select = pd.read_pickle("/Users/Tianwang/ironhack/git/03_project_ironhack_tableau/data/baton_rouge_animal_control_incidents_clean.pkl")

df_select.head()

## First returning a new dataframe with only rows, where service_code is not NaN

new = df_select.loc[df_select["service_code"].notna()]

len(new) 

len(df_select) #the old df hat 72924 rows.

to_be_dropped = len(df_select) - len(test)
to_be_dropped 

#20193 rows with emtpy Service Code will be dropped.

## Cross-check how the dropped rows look like. 

isna = df[df_select["service_code"].isna()]

isna["year"].value_counts() #Just to confirm that the insa values almost only occured in 2016-2018. 

## Create a new file with the new dataframe.

pd.to_pickle (new, "/Users/Tianwang/ironhack/git/03_project_ironhack_tableau/data/baton_rouge_animal_control_incidents_clean_update.pkl")

new.to_csv("/Users/Tianwang/ironhack/git/03_project_ironhack_tableau/data/baton_rouge_animal_control_incidents_clean_update.csv", sep=";", index=False)
