==> workbook/03. Geocoding.ipynb <==
# 3. Geocoding

## Trying to use the openCage API to translate Adress into Geodates.

import pandas as pd
import datetime
import requests
import json
import geocoder

df = pd.read_pickle("/Users/Tianwang/ironhack/git/03_project_ironhack_tableau/data/baton_rouge_animal_control_incidents_clean_update.pkl")

len(df)

location = pd.DataFrame(df["clean_adress"])

location

adress = location.drop_duplicates(subset="clean_adress", inplace=False)

adress_list = list(adress.loc[adress["clean_adress"].notna()]["clean_adress"])

len(adress_list)

### Test Geocoder to fetch the Geo coordinates using OpenCage API.

url = "https://api.opencagedata.com/geocode/v1/json?parameters"

g = geocoder.opencage("3671 N FOSTER, BATON ROUGE, 70805", key="x")

type(g.json)

#### Testing with one adress

g.json

result = g.json.keys()

key = list(result)[0]

g.json[key]

#### Test on a bigger sample

sample = adress_list[0:20]

sample

geo_list_test = []
for adress in sample:
    g = geocoder.opencage(adress, key="x")
    result = g.json.keys()
    key = list(result)[0]
    geo = g.json[key]
    geo_list_test.append(geo)

geo_list

pd.DataFrame(geo_list)

#### apply to all

first_list = adress_list[100:140]

geo_list = []
for adress in first_list:
    g = geocoder.opencage(adress, key="x")
    result = g.json.keys()
    key = list(result)[0]
    geo = g.json[key]
    geo_list.append(geo)

geo_list

### --> It worked out. However, the API limit for free version is set to 2.500 per days. As the adress list has more than 30.000 lines, I decided not to add it to my Tableau. 

## Trying to use the Bing Geo API

url = "http://spatial.virtualearth.net/REST/v1/Dataflows/Geocode?input=csv&key=x"

len(adress_list)

bing_sample = adress_list[0:50]

BingMapsAPIKey = "x"

### Due to time constraints, I will not use the API for this project.
